Introduce intelligent Fuel Station detector.Read->
Introduce some functions which checks if a fuel station exists nearby
the agv.if so,a dialog asks for refuel or not.

Algorythm is programmed to stop the animation while refuelling and
continue the animation,when refuel ends.

Although ,for the moment,it is only coded to detect fuel station which
is on the RIGHT side of AGV